\name{ZUMAppliances-package}
\alias{ZUMAppliances-package}
\alias{ZUMAppliances}
\docType{package}
\title{
\packageTitle{ZUMAppliances}
}
\description{
\packageDescription{ZUMAppliances}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ZUMAppliances}
\packageIndices{ZUMAppliances}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{ZUMAppliances}

Maintainer: \packageMaintainer{ZUMAppliances}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
